+++ inst/include/dplyr/get_column.h
@@ -3,7 +3,7 @@
 
 namespace dplyr {
 
-  Symbol get_column(SEXP, const Environment&, const ILazySubsets&);
+  SymbolString get_column(SEXP, const Environment&, const ILazySubsets&);
 
 }