+++ tests/testthat/test-distinct.R
@@ -46,3 +46,10 @@ test_that("unless .keep_all = TRUE", {
   expect_equal(df %>% distinct(x), data_frame(x = 1))
   expect_equal(df %>% distinct(x, .keep_all = TRUE), data_frame(x = 1, y = 3L))
 })
+
+test_that("distinct doesn't duplicate columns", {
+  df <- tibble(a = 1:3, b = 4:6)
+
+  expect_named(df %>% distinct(a, a), "a")
+  expect_named(df %>% group_by(a) %>% distinct(a), "a")
+})